Wired services in order to check for existent objects

+++ src/main/java/main/lab1/services/NotificationService.java
@@ -1,6 +1,6 @@
 package main.lab1.services;
 
-import main.lab1.entities.Notification;
+import main.lab1.model.Notification;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -9,9 +9,9 @@ import java.util.List;
 public interface NotificationService {
     List<Notification> getAllNotifications();
 
-    List<Notification> getNotificationsByUserId(long userId);
+    List<Notification> getNotificationsByUserId(int userId);
 
-    List<Notification> getNotificationsByTaskId(long userId);
+    List<Notification> getNotificationsByTaskId(int userId);
 
     void createNotification(Notification notification);
 }
